Вынес работу с данными в интерактор
