+++ lib/data/repository/branch_repository.dart
@@ -1,9 +1,7 @@
 
 
 import 'package:flutter_internship_v2/data/database/db_storage/branch_db_storage.dart';
-import 'package:flutter_internship_v2/data/models/all_branch_info.dart';
 import 'package:flutter_internship_v2/data/models/branch.dart';
-import 'package:flutter_internship_v2/data/models/one_branch_info.dart';
 import 'package:flutter_internship_v2/data/storage/branch_wrapper.dart';
 
 class BranchRepository{
@@ -25,12 +23,8 @@ class BranchRepository{
     branchWrapper.deleteBranch(branchID);
   }
 
-  AllBranchesInfo getAllBranchesTasksInfo() {
-    return branchWrapper.getAllBranchesTasksInfo();
-  }
-
-  List<OneBranchInfo> getAllBranchesInfo() {
-    return branchWrapper.getAllBranchesInfo();
+  Map<String, Branch> getAllBranches(){
+    return branchWrapper.getAllBranches();
   }
