+++ lib/domain/models/all_branch_info.dart
