+++ lib/domain/models/one_branch_info.dart
