+++ lib/presentation/views/branches_page/all_branch_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_internship_v2/data/models/all_branch_info.dart';
+import 'package:flutter_internship_v2/domain/models/all_branch_info.dart';
 import 'package:flutter_internship_v2/presentation/constants/my_images.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'horizontal_progress_bar.dart';
