+++ lib/presentation/views/branches_page/one_branch_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_internship_v2/data/models/one_branch_info.dart';
+import 'package:flutter_internship_v2/domain/models/one_branch_info.dart';
 import 'package:flutter_internship_v2/presentation/cubit/branch/branch_cubit.dart';
 import 'package:flutter_internship_v2/presentation/pages/all_tasks.dart';
 import 'one_branch_progress_bar.dart';
